COMP-13 Added custom exception handler and extends some of basic tests

+++ src/test/java/pl/com/company/controller/EmployeeControllerIntegrationTest.java
@@ -8,6 +8,7 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
@@ -15,6 +16,7 @@ import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.RequestBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
+import pl.com.company.exception.EmployeeException;
 import pl.com.company.service.EmployeeService;
 
 import java.math.BigDecimal;
@@ -57,13 +59,15 @@ class EmployeeControllerIntegrationTest {
 
     @AfterEach
     public void clear() {
-        this.employeeService.delete(PESEL_ONE);
-        this.employeeDto = null;
+        if (null != this.employeeService.get(PESEL_ONE)) {
+            this.employeeService.delete(PESEL_ONE);
+            this.employeeDto = null;
+        }
     }
 
     @Test
     void get() throws Exception {
-        MvcResult result = sendRequest(MockMvcRequestBuilders.get("/employee/" + PESEL_ONE).contentType(MediaType.APPLICATION_JSON));
+        MvcResult result = sendRequest(MockMvcRequestBuilders.get("/employee/" + PESEL_ONE).contentType(MediaType.APPLICATION_JSON), HttpStatus.OK);
 
         EmployeeDto employeeDtoResponse = objectMapper.readValue(result.getResponse().getContentAsString(), EmployeeDto.class);
 
@@ -77,7 +81,7 @@ class EmployeeControllerIntegrationTest {
     @Test
     void create() throws Exception {
         String employeeDtoAsJson = objectMapper.writeValueAsString(new EmployeeDto(FIRST_NAME_TWO, LAST_NAME_TWO, PESEL_TWO, SALARY_TWO));
-        MvcResult result = sendRequest(MockMvcRequestBuilders.post("/employee").content(employeeDtoAsJson).contentType(MediaType.APPLICATION_JSON));
+        MvcResult result = sendRequest(MockMvcRequestBuilders.post("/employee").content(employeeDtoAsJson).contentType(MediaType.APPLICATION_JSON), HttpStatus.OK);
 
         EmployeeDto employeeDtoResponse = objectMapper.readValue(result.getResponse().getContentAsString(), EmployeeDto.class);
 
@@ -89,9 +93,18 @@ class EmployeeControllerIntegrationTest {
     }
 
     @Test
+    void cannotCreateSecondEmployeeWithExistingPesel() throws Exception {
+        String employeeDtoAsJson = objectMapper.writeValueAsString(new EmployeeDto(FIRST_NAME_TWO, LAST_NAME_TWO, PESEL_ONE, SALARY_TWO));
+        MvcResult result = sendRequest(MockMvcRequestBuilders.post("/employee").content(employeeDtoAsJson).contentType(MediaType.APPLICATION_JSON), HttpStatus.CONFLICT);
+
+        EmployeeException exceptionDtoResponse = objectMapper.readValue(result.getResponse().getContentAsString(), EmployeeException.class);
+        assertEquals("Pesel already exist " + PESEL_ONE, exceptionDtoResponse.getMessage());
+    }
+
+    @Test
     void update() throws Exception {
         String employeeDtoAsJson = objectMapper.writeValueAsString(new EmployeeDto(FIRST_NAME_TWO, LAST_NAME_TWO, PESEL_ONE, SALARY_TWO));
-        MvcResult result = sendRequest(MockMvcRequestBuilders.put("/employee").content(employeeDtoAsJson).contentType(MediaType.APPLICATION_JSON));
+        MvcResult result = sendRequest(MockMvcRequestBuilders.put("/employee").content(employeeDtoAsJson).contentType(MediaType.APPLICATION_JSON), HttpStatus.OK);
 
         EmployeeDto employeeDtoResponse = objectMapper.readValue(result.getResponse().getContentAsString(), EmployeeDto.class);
 
@@ -104,16 +117,27 @@ class EmployeeControllerIntegrationTest {
 
     @Test
     void delete() throws Exception {
-        MvcResult result = sendRequest(MockMvcRequestBuilders.delete("/employee/{pesel}", PESEL_ONE).contentType(MediaType.APPLICATION_JSON));
+        MvcResult result = sendRequest(MockMvcRequestBuilders.delete("/employee/{pesel}", PESEL_ONE).contentType(MediaType.APPLICATION_JSON), HttpStatus.OK);
 
         boolean isDeleted = objectMapper.readValue(result.getResponse().getContentAsString(), Boolean.class);
 
         assertTrue(isDeleted);
     }
 
-    private MvcResult sendRequest(RequestBuilder request) throws Exception {
+    @Test
+    void cannotDeleteNotExistingEmployee() throws Exception {
+        sendRequest(MockMvcRequestBuilders.delete("/employee/{pesel}", PESEL_ONE).contentType(MediaType.APPLICATION_JSON), HttpStatus.OK);
+
+        MvcResult result = sendRequest(MockMvcRequestBuilders.delete("/employee/{pesel}", PESEL_ONE).contentType(MediaType.APPLICATION_JSON), HttpStatus.CONFLICT);
+        EmployeeException employeeDtoResponse = objectMapper.readValue(result.getResponse().getContentAsString(), EmployeeException.class);
+
+        assertEquals("Employee with pesel: " + PESEL_ONE + " does not exist", employeeDtoResponse.getMessage());
+
+    }
+
+    private MvcResult sendRequest(RequestBuilder request, HttpStatus expectedStatus) throws Exception {
         return mockMvc.perform(request)
-                .andExpect(MockMvcResultMatchers.status().is(200))
+                .andExpect(MockMvcResultMatchers.status().is(expectedStatus.value()))
                 .andExpect(MockMvcResultMatchers.content()
                         .contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
                 .andReturn();
